+++ django_apscheduler/result_storage.py
@@ -7,19 +7,21 @@ from django_apscheduler.models import DjangoJobExecution, DjangoJob
 from django_apscheduler.util import serialize_dt
 
 
-class DjangoResultStorage(object):
+class DjangoResultStorage:
     """
     Uses Django ORM table for store job status and results.
+
     You can override this class to change result storage.
     """
 
-    LOGGER = logging.getLogger("django_apscheduler.result_storage")
+    logger = logging.getLogger("django_apscheduler.result_storage")
 
     def get_or_create_job_execution(
         self, job: DjangoJob, event: JobSubmissionEvent
     ) -> int:
         """
         Create and return new job execution item.
+
         :param job: DjangoJob instance
         :type job: django_apscheduler.models.DjangoJob
         :param event: JobSubmissionEvent instance
@@ -54,11 +56,10 @@ class DjangoResultStorage(object):
             run_time=serialize_dt(event.scheduled_run_times[0]),
         ).id
 
-    def register_job_executed(
-        self, job: DjangoJobExecution, event: JobExecutionEvent
-    ) -> int:
+    def register_job_executed(self, job: DjangoJobExecution, event: JobExecutionEvent):
         """
-        Registration of job execution status
+        Registration of job execution status.
+
         :param job: DjangoJob instance
         :param event: JobExecutionEvent instance
         :return: JobExecution id
@@ -71,7 +72,7 @@ class DjangoResultStorage(object):
             )
             .order_by("id")
             .last()
-        )  # type: DjangoJobExecution
+        )
 
         if not job_execution:
             job_execution = DjangoJobExecution.objects.create(
@@ -81,7 +82,7 @@ class DjangoResultStorage(object):
             )
 
         if job_execution.finished:
-            self.LOGGER.warning("Job already finished! %s", job_execution)
+            self.logger.warning("Job already finished! %s", job_execution)
             return
 
         job_execution.finished = time.time()