+++ django_apscheduler/util.py
@@ -5,18 +5,21 @@ from django.utils import timezone
 
 def serialize_dt(dt):
     """
-    Converts datetime with timezone to datetime without timezone if USE_TZ is False
+    Converts datetime with timezone to datetime without timezone if USE_TZ is False.
+
     :param dt:
     :return:
     """
     if not settings.USE_TZ and dt and timezone.is_aware(dt):
         return timezone.make_naive(dt)
+
     return dt
 
 
 def deserialize_dt(dt):
     if not settings.USE_TZ and dt and timezone.is_naive(dt):
         return timezone.make_aware(dt)
+
     return dt